feat: Console 클래스 위치 변경 및 바우처 정보 출력 구현
